Split `Common.kt` into multiple files in a new `common/` package
