+++ app/src/androidTest/java/com/github/se/signify/ui/CommonTest.kt
@@ -26,6 +26,23 @@ import com.github.se.signify.R
 import com.github.se.signify.model.di.AppDependencyProvider
 import com.github.se.signify.model.hand.HandLandMarkViewModel
 import com.github.se.signify.model.navigation.NavigationActions
+import com.github.se.signify.ui.common.AccountInformation
+import com.github.se.signify.ui.common.AnnexScreenScaffold
+import com.github.se.signify.ui.common.BackButton
+import com.github.se.signify.ui.common.BasicButton
+import com.github.se.signify.ui.common.CameraBox
+import com.github.se.signify.ui.common.InfoPopup
+import com.github.se.signify.ui.common.LearnedLetterList
+import com.github.se.signify.ui.common.LetterList
+import com.github.se.signify.ui.common.MainScreenScaffold
+import com.github.se.signify.ui.common.NotImplementedYet
+import com.github.se.signify.ui.common.ProfilePicture
+import com.github.se.signify.ui.common.ScreenColumn
+import com.github.se.signify.ui.common.SquareButton
+import com.github.se.signify.ui.common.StatisticsTable
+import com.github.se.signify.ui.common.StreakCounter
+import com.github.se.signify.ui.common.TextButton
+import com.github.se.signify.ui.common.TopBar
 import org.junit.Rule
 import org.junit.Test
 import org.mockito.Mockito.mock
