+++ app/src/androidTest/java/com/github/se/signify/ui/screens/profile/SettingsScreenTest.kt
@@ -16,7 +16,7 @@ import com.github.se.signify.model.auth.UserSession
 import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.user.UserRepository
 import com.github.se.signify.model.user.UserViewModel
-import com.github.se.signify.ui.ProfilePicture
+import com.github.se.signify.ui.common.ProfilePicture
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
