+++ app/src/main/java/com/github/se/signify/ui/screens/challenge/ChallengeScreen.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.unit.dp
 import com.github.se.signify.R
 import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.navigation.Screen
-import com.github.se.signify.ui.MainScreenScaffold
-import com.github.se.signify.ui.SquareButton
+import com.github.se.signify.ui.common.MainScreenScaffold
+import com.github.se.signify.ui.common.SquareButton
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
