+++ app/src/main/java/com/github/se/signify/ui/screens/challenge/NewChallengeScreen.kt
@@ -45,8 +45,8 @@ import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.navigation.Screen
 import com.github.se.signify.model.user.UserRepository
 import com.github.se.signify.model.user.UserViewModel
-import com.github.se.signify.ui.AnnexScreenScaffold
-import com.github.se.signify.ui.TextButton
+import com.github.se.signify.ui.common.AnnexScreenScaffold
+import com.github.se.signify.ui.common.TextButton
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
