+++ app/src/main/java/com/github/se/signify/ui/screens/home/ExerciseScreen.kt
@@ -44,8 +44,8 @@ import com.github.se.signify.model.hand.HandLandMarkViewModel
 import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.stats.StatsRepository
 import com.github.se.signify.model.stats.StatsViewModel
-import com.github.se.signify.ui.AnnexScreenScaffold
-import com.github.se.signify.ui.CameraBox
+import com.github.se.signify.ui.common.AnnexScreenScaffold
+import com.github.se.signify.ui.common.CameraBox
 
 /**
  * Composable function for a common exercise screen layout, handling gesture detection, sentence
