+++ app/src/main/java/com/github/se/signify/ui/screens/home/FeedbackScreen.kt
@@ -43,8 +43,8 @@ import com.github.se.signify.model.feedback.FeedbackRepository
 import com.github.se.signify.model.feedback.FeedbackViewModel
 import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.navigation.Screen
-import com.github.se.signify.ui.AnnexScreenScaffold
-import com.github.se.signify.ui.TextButton
+import com.github.se.signify.ui.common.AnnexScreenScaffold
+import com.github.se.signify.ui.common.TextButton
 
 @Composable
 fun FeedbackScreen(
