+++ app/src/main/java/com/github/se/signify/ui/screens/home/QuizScreen.kt
@@ -43,7 +43,7 @@ import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.quiz.QuizQuestion
 import com.github.se.signify.model.quiz.QuizRepository
 import com.github.se.signify.model.quiz.QuizViewModel
-import com.github.se.signify.ui.AnnexScreenScaffold
+import com.github.se.signify.ui.common.AnnexScreenScaffold
 
 @Composable
 fun QuizScreen(navigationActions: NavigationActions, quizRepository: QuizRepository) {
