+++ app/src/main/java/com/github/se/signify/ui/screens/profile/FriendsListScreen.kt
@@ -61,8 +61,8 @@ import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.model.navigation.Screen
 import com.github.se.signify.model.user.UserRepository
 import com.github.se.signify.model.user.UserViewModel
-import com.github.se.signify.ui.AnnexScreenScaffold
-import com.github.se.signify.ui.ProfilePicture
+import com.github.se.signify.ui.common.AnnexScreenScaffold
+import com.github.se.signify.ui.common.ProfilePicture
 import kotlinx.coroutines.delay
 
 @Composable
